ZipcodeNewBuffer update
Fixed the problem with not being able to overwrite the file. Now if the program wants to rewrite the header line it copies the saved header line and copies it and the rest of the data to a new file, deletes the original file, and then renames the copy file to the same name as the original file.

(Kind of a round about method of doing it, but I couldn't figure out how to directly edit the file)